requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.1
attrs==21.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
comtypes==1.1.8
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mediapipe==0.8.5
mysql-connector==2.2.9
mysql-connector-python==8.0.23
numpy==1.20.1
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
panda==0.3.1
pandas==1.2.4
pdfkit==0.6.1
Pillow==8.2.0
protobuf==3.15.5
psycopg2==2.9.1
pypiwin32==223
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
requests==2.25.1
scipy==1.6.3
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.11
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
whitenoise==5.2.0
wikipedia==1.4.0
wkhtmltopdf==0.2
WTForms==2.3.3
